Replace ES2015 by ES5 syntax
This commit replace some line of code using the arrow functions of
ES2015. The reason is, for now, Safari does not support this feature.
Using a transpiler like Babel and Traceur could fix the issue but it
seems not so easy to make it compatible with others dependencies
(momentjs does not support the transpile step for example) and I think
it's not the responsibility of the "paper-datatable" useres to do this.
The element should be bundled.